% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webRmd.R
\name{convertWebRmd2drakeplanAndMake}
\alias{convertWebRmd2drakeplanAndMake}
\title{Convert webRmd 2 drake plan script and make}
\usage{
convertWebRmd2drakeplanAndMake()
}
\value{
A list of planDetails
}
\description{
Convert webRmd 2 drake plan script and make
}
\examples{
none
}
