% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{create_reactiveScript}
\alias{create_reactiveScript}
\title{Create Reactive script from drake}
\usage{
create_reactiveScript()
}
\value{

}
\description{
Create Reactive script from drake
}
\examples{
None
}
