% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeRmd2drakeRmd.R
\name{generate_drakeRmdFromActiveRmd}
\alias{generate_drakeRmdFromActiveRmd}
\title{Generate drake Rmd from active Rmd (for bookdown writing)}
\usage{
generate_drakeRmdFromActiveRmd(
  needMakeconditionHolder = F,
  removeLabelSuffix = "_[0-9]{2}"
)
}
\arguments{
\item{needMakeconditionHolder}{A logical default=F meaning no needs for a makecondition code chunk prepared for you.}

\item{removeLabelSuffix}{A character to represent "{removeLabelSuffix}$" regex pattern, default = "_[0-9]{2}". If not intend for removal, set it to "".}
}
\value{

}
\description{
Generate drake Rmd from active Rmd (for bookdown writing)
}
\examples{
none
}
