% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandplan.R, R/plan_details.R
\name{generateGrandplanScriptFromActiveRmd}
\alias{generateGrandplanScriptFromActiveRmd}
\title{Generate grandplan_XXX.R from active Rmd main plan}
\usage{
generateGrandplanScriptFromActiveRmd()

generateGrandplanScriptFromActiveRmd()
}
\value{
A list of grandplanDetails

A list of grandplanDetails
}
\description{
Generate grandplan_XXX.R from active Rmd main plan

Generate grandplan_XXX.R from active Rmd main plan
}
\examples{
none
none
}
