% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{get_chunksTable}
\alias{get_chunksTable}
\title{Get chunk labels start end position in Rmdlines}
\usage{
get_chunksTable(
  Rmdlines,
  engine = "r",
  requireLabel = F,
  exclude = "(afterMake=T|drake=F|\\\\bsetup\\\\b)"
)
}
\arguments{
\item{Rmdlines}{character. lines extracted from readLines}

\item{engine}{character. default="r"}

\item{requireLabel}{logical. default=T. extracted only those chunks with a labels.}

\item{exclude}{character. Regex represents pattern in starting ticks. Match will be ignored.}
}
\value{
A data frame table
}
\description{
Get chunk labels start end position in Rmdlines
}
\examples{
None.
}
