% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{get_listCodeChunksFromRmdlines}
\alias{get_listCodeChunksFromRmdlines}
\title{Get a list of code chunks from Rmdlines}
\usage{
get_listCodeChunksFromRmdlines(
  Rmdlines,
  engine = "r",
  stackNAchunks = T,
  requireLabel = F,
  exclude = "(afterMake=T|drake=F|\\\\bsetup\\\\b)"
)
}
\arguments{
\item{Rmdlines}{A character string from readLines.}

\item{engine}{A character specifies which language code chunk to target}

\item{stackNAchunks}{A logical. default=T the unlabelled code chunks will be stack together}

\item{requireLabel}{A logical. default=F extract unlabelled chunks as well}

\item{exclude}{A character of regex setting labels pattern whose chunks will be ignored from extraction.}
}
\value{

}
\description{
Get a list of code chunks from Rmdlines
}
\examples{
None.
}
