% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandplan.R
\name{mk_currentActiveGrandPlan}
\alias{mk_currentActiveGrandPlan}
\title{make function for current active grandplan}
\usage{
mk_currentActiveGrandPlan()
}
\value{

}
\description{
make function for current active grandplan
}
\examples{
none
}
