% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{mkActiveRmd_plan}
\alias{mkActiveRmd_plan}
\title{make active Rmd's "(grand)plan_{activeRmdName}"}
\usage{
mkActiveRmd_plan()
}
\value{

}
\description{
make active Rmd's "(grand)plan_{activeRmdName}"
}
\examples{
purlActiveRmd_thenPlanMake()
}
