% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake2rmd.R
\name{produceRmdFromActiveRscript}
\alias{produceRmdFromActiveRscript}
\title{Generate drake plan Rmd from active drake plan R script}
\usage{
produceRmdFromActiveRscript(removePlanPrefix = T)
}
\arguments{
\item{removePlanPrefix}{A logical. default=T will remove 'plan_' from R script filename}
}
\value{

}
\description{
Generate drake plan Rmd from active drake plan R script
}
\examples{
none
}
