% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandplan.R
\name{purlActiveMainRmd_thenPlanMake}
\alias{purlActiveMainRmd_thenPlanMake}
\title{Purl Activive Main Plan Rmd, with its subplan Rmds to a Grandplan
then make it}
\usage{
purlActiveMainRmd_thenPlanMake(visAfterMake = T)
}
\value{

}
\description{
Purl Activive Main Plan Rmd, with its subplan Rmds to a Grandplan
then make it
}
\examples{
none
}
