% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{set_gitignore}
\alias{set_gitignore}
\title{Setup .gitignore to fit R users, also purge those tracked-but-should-be-ignored files}
\usage{
set_gitignore()
}
\value{

}
\description{
Setup .gitignore to fit R users, also purge those tracked-but-should-be-ignored files
}
\examples{
none
}
