% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandplan.R
\name{vis_currentActiveMainPlan}
\alias{vis_currentActiveMainPlan}
\title{Visualize the current active grand plan}
\usage{
vis_currentActiveMainPlan()
}
\value{

}
\description{
Visualize the current active grand plan
}
\examples{
none
}
