% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{visActiveRmd_plan}
\alias{visActiveRmd_plan}
\title{visualize active Rmd's "(grand)plan_{activeRmdName}"}
\usage{
visActiveRmd_plan()
}
\value{

}
\description{
visualize active Rmd's "(grand)plan_{activeRmdName}"
}
\examples{
purlActiveRmd_thenPlanMake()
}
