% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_switch.R
\name{convert2drakeplanAndMake}
\alias{convert2drakeplanAndMake}
\title{Convert 2 drake plan script and make}
\usage{
convert2drakeplanAndMake()
}
\value{
A list of planDetails
}
\description{
Convert 2 drake plan script and make
}
\examples{
none
}
