% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplan.R
\name{purlActiveMainRmd_thenPlanMake}
\alias{purlActiveMainRmd_thenPlanMake}
\title{purl the main Rmd in multiple subplan Rmd cases}
\usage{
purlActiveMainRmd_thenPlanMake()
}
\value{

}
\description{
purl the main Rmd in multiple subplan Rmd cases
}
\examples{
none
}
